src/main/kotlin/no/nav/dagpenger/soknad/orkestrator/behov/løsere/HelseTilAlleTyperJobbBehovløser.kt
package no.nav.dagpenger.soknad.orkestrator.behov.løsere

import no.nav.dagpenger.soknad.orkestrator.behov.Behovløser
import no.nav.dagpenger.soknad.orkestrator.behov.BehovløserFactory.Behov.HelseTilAlleTyperJobb
import no.nav.dagpenger.soknad.orkestrator.quizOpplysning.db.QuizOpplysningRepository
import no.nav.helse.rapids_rivers.RapidsConnection

class HelseTilAlleTyperJobbBehovløser(
    rapidsConnection: RapidsConnection,
    opplysningRepository: QuizOpplysningRepository,
) : Behovløser(rapidsConnection, opplysningRepository) {
    override val behov = HelseTilAlleTyperJobb.name
    override val beskrivendeId = "faktum.alle-typer-arbeid"
}
